Add files via upload
Ask the students not in your pod group what their favorite candy to get on Halloween.

Compile the list of all the candy. Print the list so that you have everyone's name and corresponding candy

Print the list.

Print each item in the list individually

After you print the list you need to sort the list by name and print it again.

Print each item in the list individually again but skip a line between each item

Add Mr. McKinstry's favorite type of candy to the list.

Print the list in reverse order.

Lastly, find and print the length of your list.

Turn in your Python file.