Add files via upload
The practical test will use the Candy project from the previous Practical Test.

Loop throught the list and create a print statement that tell me this is the candy your group would like.

Create a second list that shows the cost of 6 of each types of candy. You will have to find that information. (Try Amazon)

Concantinate the two list to show the cost of each type of candy. If you can do this as a loop that would be great but not necessary.

Create a copy of the candy list and call it alternate candy. You will append the list and add two other types of candy that we can order if the second and third types of candy in your list are not available.

Take the second and third types of candy out and create a print statement using your list that will say if "item 2 is not available the replace it with one of the new types you entered into you list. Do this again for the third item in you list. Make sure you alternate items are not the same.

At the end of your program print candy_list and alternate_candy. Hint they should be different