Add files via upload
Using the Python file

Create a Dictionary from the two lists where candy_type is the key and candy_price is the value.

Check for Duplicate candy types

Remove any candy that costs more than $3.00

Add a new value to the dictionary and that tells me if the candy is chocolate or some other type of candy. (You might have to research this and manually enter this.

Print out all candy that is chocolate and remind me to make sure to keep refrigerated.

In a docstring explain why it is easier to use the candy data in a dictionary than it was in a list. Also, explain when and why we use lists and dictionaries.